+++ packages/core/src/components/responsive-element/responsive-element.template.ts
@@ -1,0 +1,10 @@
+import { ResponsiveElement } from '@dso-toolkit/sources';
+import { html, TemplateResult } from 'lit-html';
+
+export function responsiveElementTemplate({ children }: ResponsiveElement<TemplateResult>) {
+  return html`
+    <dso-responsive-element>
+      ${children}
+    </dso-responsive-element>
+  `;
+}