+++ packages/css/src/components/tile-grid/tile-grid.template.ts
@@ -3,8 +3,8 @@ import { html, TemplateResult } from 'lit-html';
 
 export function tileGridTemplate({ children }: TileGrid<TemplateResult>) {
   return html`
-    <div class="dso-tile-grid">
+    <dso-responsive-element class="dso-tile-grid">
       ${children}
-    </div>
+    </dso-responsive-element>
   `;
 }