+++ packages/sources/src/components/document-header/document-header.args.ts
@@ -1,0 +1,31 @@
+import { HandlerFunction } from '@storybook/addon-actions';
+import { DefinitionList } from '../definition-list/definition-list.models';
+import { DocumentHeader } from './document-header.models';
+
+export interface DocumentHeaderArgs {
+  title: string;
+  owner: string;
+  type: string;
+  featureAction: HandlerFunction;
+  featuresOpen: boolean;
+  statusContentOpen: boolean;
+}
+
+export function documentHeaderMapper<TemplateFnReturnType>(
+  a: DocumentHeaderArgs,
+  status: TemplateFnReturnType,
+  features: DefinitionList<TemplateFnReturnType>,
+  statusContent: TemplateFnReturnType
+): DocumentHeader<TemplateFnReturnType> {
+  return {
+    title: a.title,
+    owner: a.owner,
+    status,
+    type: a.type,
+    features,
+    featureAction: a.featureAction,
+    featuresOpen: a.featuresOpen,
+    statusContent,
+    statusContentOpen: a.statusContentOpen
+  };
+}