+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Deprecated
 * **BREAKING: dso-toolkit:** Conclusion ([#1624](https://github.com/dso-toolkit/dso-toolkit/issues/1624))
 
+### Added
+* **core:** Nieuw component `<dso-responsive-element />` ([#1456](https://github.com/dso-toolkit/dso-toolkit/issues/1456))
+
 ### Changed
 * **css + sources:** Accordion button en link implementatie etaleren ([#1481](https://github.com/dso-toolkit/dso-toolkit/issues/1481))
 * **dso-toolkit + css:** Form Group: Confirm naar @dso-toolkit/css ([#1611](https://github.com/dso-toolkit/dso-toolkit/issues/1611))